requirements.txt
python-telegram-bot==21.6
aiosqlite==0.20.0
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
fastapi==0.115.5
uvicorn==0.32.1
jinja2==3.1.4
python-multipart==0.0.17
